% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_technical_coefficients.R
\name{calculate_technical_coefficients}
\alias{calculate_technical_coefficients}
\title{Calculate Technical Coefficients}
\usage{
calculate_technical_coefficients(data, total_output)
}
\arguments{
\item{data}{A data frame containing the inter-industry transactions data.}

\item{total_output}{A numeric vector containing total output for each industry.}
}
\value{
A numeric matrix with the technical coefficients.
}
\description{
This function converts a specified range of a data frame to a
numeric matrix and divides each column by the corresponding value in the
total_output vector.
}
\details{
The function first converts the specified range of the data frame
to a numeric matrix. It then uses the `sweep` function to divide each column
by the corresponding value in the `total_output` vector, resulting in a
matrix of technical coefficients.
}
